Объединил компоненты edit-event и new-event в один - new-edit-event
